Move class MV to galgebra.deprecated
